Add files via upload
We are making a hardware which can be converted into a band which sense the body temperature and the position of the hand and further we can make it to sense the further characteristics of the body which are used to detect the symptoms of the corona. The device consist of ESP32, temperature sensor(LM35), MPU6050. The ESP32 is a microcontroller having inbuilt bluetooth which is used to connect our device with the phone. MPU6050 is used to get the position of the hand. By checking the value of the angles when our hand comes near to the face the inbuilt led of esp32 starts glowing. The temperature sensor sense the body temperature sends the value to the phone and as soon as the temperature goes above the 102 Fahrenheit then the device sends an alert to the phone.